Changes to use ConsumerRebal..ner
