+++ data-plane/dispatcher/src/main/java/dev/knative/eventing/kafka/broker/dispatcher/impl/consumer/UnorderedConsumerVerticle.java
@@ -59,9 +59,9 @@ public final class UnorderedConsumerVerticle extends ConsumerVerticle {
 
   @Override
   void startConsumer(final Promise<Void> startPromise) {
-    Objects.requireNonNull(getConsumerVerticleContext().getConsumerRebalanceListener());
+    Objects.requireNonNull(getConsumerRebalanceListener());
 
-    this.consumer.subscribe(Set.copyOf(getConsumerVerticleContext().getResource().getTopicsList()), getConsumerVerticleContext().getConsumerRebalanceListener())
+    this.consumer.subscribe(Set.copyOf(getConsumerVerticleContext().getResource().getTopicsList()), (getConsumerRebalanceListener()))
     .onFailure(startPromise::tryFail)
     .onSuccess(startPromise::tryComplete)
     .onSuccess(v -> poll());
